+++ lib/milo/product.rb
@@ -1,30 +1,33 @@
 module Milo
   module Product
 
-    def get_products(options = {})
-      base = "products"
-      apply_flag(base, "?", options[:show])
-      make_request(base)
-    end
-
     def get_product_by_id(id, options = {})
       base = "products?product_ids=#{id}"
-      apply_flag(base, "&", options[:show])
-      make_request(base)
+      result = apply_flag(base, options)
+      make_request(result)
     end
 
     def get_product_by_upc(upc, options = {})
       base = "products?q=upc:#{upc}"
-      apply_flag(base, "&", options[:show])
-      make_request(base)
+      result = apply_flag(base, options)
+      make_request(result)
+    end
+
+    def get_product_by_postal_code(code, options = {})
+      base = "products?postal_code=#{code}"
+      result = apply_flag(base, options)
+      make_request(result)
     end
 
     private
 
-    def apply_flag(base, identifier, flag)
-      base.tap do |b|
-        b.concat identifier.concat "show=#{flag}" if flag.present?
-      end
+    def apply_flag(base, options = {})
+      return base if options == {}
+      hash = {}
+      hash.merge!(show: options[:show]) if options[:show]
+      hash.merge!(q: options[:q]) if options[:q]
+      hash.merge!(show_defaults: options[:show_defaults]) if options[:show_defaults]
+      base + "&" + hash.to_query
     end
   end
 end