+++ spec/milo/product_spec.rb
@@ -1,48 +1,43 @@
 # encoding: utf-8
 require "milo"
+require 'spec_helper'
 
 describe Milo::Main, 'product api' do
 
   before :each do
-    @key = "my_key"
-    curl_object = Curl::Easy.method(:new)
-    Curl::Easy.stub!(:new).and_return do |*args|
-      curl = curl_object.call(*args)
-      curl.stub!(:perform)
-      curl.stub!(:header_str) { "\r\ntest: blah"}
-      curl.stub!(:response_code).and_return(200)
-      curl.stub!(:body_str).and_return("{\"pagination\": {\"total_results\": 1, \"per_page\": 30, \"total_pages\": 1, \"page\": 0}, \"products\": [{\"merchants\": [4483], \"min_price\": null, \"name\": \"REDI2SET 43\\\" x 14\\\" Wavy Glass Pattern Frameless Replacement Glass Block Window\", \"max_price\": null, \"product_id\": 20482374}]}")
-      curl
-    end
+    @key = "ffbefff2a246f4c53c0d04ed50bb4707"
   end
 
-  it "should be able to list all" do
-    milo = Milo::Main.new(@key)
-    response = milo.get_products
-    response.url.should == "https://api.x.com/milo/v3/products?key=#{@key}"
-    response.status.should == 200
-  end
-
-  it "should be able to get by id" do
+  it "should be able to get by id", :vcr do
     milo = Milo::Main.new(@key)
     response = milo.get_product_by_id("20482374")
-    response.url.should == "https://api.x.com/milo/v3/products?product_ids=20482374&key=#{@key}"
-    response.body.should =~ /Wavy Glass Pattern Frameless Replacement Glass Block Window/
-    response.status.should == 200
+    response.to_json.should have_json_size(1).at_path("products")
   end
 
-  it "should be able to get by upc" do
+  it "should be able to get by upc", :vcr do
     milo = Milo::Main.new(@key)
     response = milo.get_product_by_upc("037000185062")
-    response.url.should == "https://api.x.com/milo/v3/products?q=upc:037000185062&key=#{@key}"
-    response.status.should == 200
+    response.to_json.should have_json_size(1).at_path("products")
   end
 
-    it "should be able to get by id with show flag in the show parameter" do
+  it "should be able to get by id with show flag in the show parameter", :vcr do
     milo = Milo::Main.new(@key)
-    response = milo.get_product_by_id("20482374", show: "PnamePminUpcImg45")
-    response.url.should == "https://api.x.com/milo/v3/products?product_ids=20482374&show=PnamePminUpcImg45&key=#{@key}"
-    response.status.should == 200
+    response = milo.get_product_by_id("20482374", {show: "PnamePminUpcImg45"})
+    response.to_json.should have_json_size(1).at_path("products")
+    response.to_json.should include_json('http://imagethumbnails.milo.com/035/340/575/45/35340237_34484575_45.jpg'.to_json).at_path("products/0/image_45")
+  end
+
+  it "should be able to get by postal code", :vcr do
+    milo = Milo::Main.new(@key)
+    response = milo.get_product_by_postal_code("07032")
+    response.to_json.should have_json_size(30).at_path("products")
+  end
+
+  it "should be able to get by postal code with many flags", :vcr do
+    milo = Milo::Main.new(@key)
+    p response = milo.get_product_by_postal_code("07032", {q: "basketball", show_defaults: "false",  show: "PidPnameMidsPminPmaxDescPurlImg200Rate"})
+    response.to_json.should have_json_size(30).at_path("products")
+    response.to_json.should include_json("Lifetime 44 in. Streamline Basketball System".to_json).at_path("products/0/name")
   end
 
 end