+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v.0.0.7
+* get_products deprecated
+* responses are parsed in JSON
+* find products by postal code
+
 ## v.0.06
 
 * get products